requirements.txt
asgiref==3.5.2
Django==4.1.4
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
djongo==1.3.6
dnspython==2.2.1
mongoengine==0.24.2
# mysqlclient==2.1.1
PyJWT==1.7.1
pymongo==3.12.3
pytz==2022.6
sqlparse==0.2.4
requests
pillow
python-dateutil
python-dotenv

# asgiref==3.5.2
# async-timeout==4.0.2
# attrs==22.1.0
# autobahn==22.7.1
# Automat==22.10.0
# cffi==1.15.1
# channels==4.0.0
# channels-redis==4.0.0
# constantly==15.1.0
# cryptography==38.0.4
# daphne==4.0.0
# hyperlink==21.0.0
# idna==3.4
# incremental==22.10.0
# msgpack==1.0.4
# psycopg2==2.9.5
# pyasn1==0.4.8
# pyasn1-modules==0.2.8
# pycparser==2.21
# pyOpenSSL==22.1.0
# redis==4.4.0
# service-identity==21.1.0
# Twisted==22.10.0
# twisted-iocpsupport==1.0.2
# txaio==22.2.1
# typing-extensions==4.4.0
# tzdata==2022.7
# zope.interface==5.5.2